app/requirements.txt
alembic==1.11.3
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pytest==7.4.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
